+++ spec/associations/device_auth/video_spec.rb
@@ -7,7 +7,7 @@ describe Yt::Video, :device_app do
   context 'given an existing video' do
     let(:id) { 'MESycYJytkU' }
 
-    it { expect(video.details_set).to be_a Yt::DetailsSet }
+    it { expect(video.content_detail).to be_a Yt::ContentDetail }
     it { expect(video.snippet).to be_a Yt::Snippet }
     it { expect(video.rating).to be_a Yt::Rating }
     it { expect(video.status).to be_a Yt::Status }
@@ -34,7 +34,7 @@ describe Yt::Video, :device_app do
   context 'given an unknown video' do
     let(:id) { 'not-a-video-id' }
 
-    it { expect{video.details_set}.to raise_error Yt::Errors::NoItems }
+    it { expect{video.content_detail}.to raise_error Yt::Errors::NoItems }
     it { expect{video.snippet}.to raise_error Yt::Errors::NoItems }
     it { expect{video.rating}.to raise_error Yt::Errors::NoItems }
     it { expect{video.status}.to raise_error Yt::Errors::NoItems }